Switch metrics from usize to u64
It's easier to convert usize values to u64 values than it is to go the
other way. Metrics are generally used as outputs rather than inputs, so
it makes sense to have it be the type that everything can easily convert
to.

+++ src/vmm/src/devices/virtio/vsock/csm/connection.rs
@@ -232,7 +232,7 @@ where
                         // On a successful data read, we fill in the packet with the RW op, and
                         // length of the read data.
                         pkt.set_op(uapi::VSOCK_OP_RW).set_len(read_cnt as u32);
-                        METRICS.vsock.rx_bytes_count.add(read_cnt);
+                        METRICS.vsock.rx_bytes_count.add(read_cnt as u64);
                     }
                     self.rx_cnt += Wrapping(pkt.len());
                     self.last_fwd_cnt_to_peer = self.fwd_cnt;
@@ -476,7 +476,7 @@ where
                     0
                 });
             self.fwd_cnt += Wrapping(flushed as u32);
-            METRICS.vsock.tx_bytes_count.add(flushed);
+            METRICS.vsock.tx_bytes_count.add(flushed as u64);
 
             // If this connection was shutting down, but is waiting to drain the TX buffer
             // before forceful termination, the wait might be over.
@@ -628,7 +628,7 @@ where
         };
         // Move the "forwarded bytes" counter ahead by how much we were able to send out.
         self.fwd_cnt += Wrapping(written as u32);
-        METRICS.vsock.tx_bytes_count.add(written);
+        METRICS.vsock.tx_bytes_count.add(written as u64);
 
         // If we couldn't write the whole slice, we'll need to push the remaining data to our
         // buffer.
